+++ src/modules/navigator/navigator.h
@@ -270,8 +270,8 @@ public:
 	bool		start_mission_landing() { return _mission.land_start(); }
 	bool		get_mission_start_land_available() { return _mission.get_land_start_available(); }
 	int 		get_mission_landing_index() { return _mission.get_land_start_index(); }
-	float 	get_mission_landing_lat() { return _mission.get_landing_lat(); }
-	float 	get_mission_landing_lon() { return _mission.get_landing_lon(); }
+	double 	get_mission_landing_lat() { return _mission.get_landing_lat(); }
+	double 	get_mission_landing_lon() { return _mission.get_landing_lon(); }
 	float 	get_mission_landing_alt() { return _mission.get_landing_alt(); }
 
 	// RTL
